refactor(cli): start imagery theme — migrate asf validator onto discovery
Set up the imagery earthlens.cli entry point (earthlens._imagery_cli) and migrate
asf's validator to earthlens.asf.cli. Relocate its tests to earthlens-imagery.
Shrink the guard allow-list (asf).
