+++ libs/core/tests/cli/test_validate.py
@@ -313,62 +313,6 @@ class TestStructuralLints:
         assert any("mars_basin" in i and "not in" in i for i in issues), "bad basin"
         assert any("jtwc" in i for i in issues), "unsupported source flagged"
 
-    def test_asf_flags_unknown_constant_names(self):
-        """An asf row whose PLATFORM/PRODUCT_TYPE constant is gone is flagged."""
-        from earthlens.cli.validate import _validate_asf
-
-        catalog = SimpleNamespace(
-            datasets={
-                "bad-row": SimpleNamespace(
-                    platform="NOT_A_PLATFORM",
-                    dataset=None,
-                    product_type="NOT_A_TYPE",
-                ),
-            }
-        )
-        _checked, issues = _validate_asf(catalog)
-        assert any("NOT_A_PLATFORM" in i for i in issues), "platform miss flagged"
-        assert any("NOT_A_TYPE" in i for i in issues), "product_type miss flagged"
-
-    def test_asf_flags_unknown_dataset_constant(self):
-        """An asf row whose DATASET constant is gone is flagged."""
-        from earthlens.cli.validate import _validate_asf
-
-        catalog = SimpleNamespace(
-            datasets={
-                "bad-row": SimpleNamespace(
-                    platform=None,
-                    dataset="NOT_A_DATASET",
-                    product_type="SLC",
-                ),
-            }
-        )
-        _checked, issues = _validate_asf(catalog)
-        assert any("NOT_A_DATASET" in i for i in issues), "dataset miss flagged"
-
-    def test_asf_reports_zero_checked_when_sdk_missing(self, monkeypatch):
-        """A missing `asf_search` returns checked=0 and an install-hint issue."""
-        import builtins
-
-        from earthlens.cli.validate import _validate_asf
-
-        real_import = builtins.__import__
-
-        def fake_import(name, *args, **kwargs):
-            if name == "asf_search":
-                raise ImportError("missing")
-            return real_import(name, *args, **kwargs)
-
-        monkeypatch.setattr(builtins, "__import__", fake_import)
-        catalog = SimpleNamespace(
-            datasets={"row-1": SimpleNamespace(), "row-2": SimpleNamespace()}
-        )
-        checked, issues = _validate_asf(catalog)
-        assert checked == 0
-        assert issues and "asf_search" in issues[0]
-        # The install hint mentions the curated row count for context.
-        assert "2 curated" in issues[0]
-
 
 class TestValidateOne:
     """Tests for validate_one."""
